requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
waitress==3.0.0
fast-agent-mcp==0.2.18
mcp==1.6.0
mcp-agent==0.0.17
uvicorn==0.30.6